backend/requirements.txt
appdirs==1.4.4
astroid==2.15.5
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
colorclass==2.2.2
dill==0.3.6
docopt==0.6.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
idna==3.4
ifaddr==0.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
lxml==4.9.3
MarkupSafe==2.1.3
mccabe==0.7.0
numpy==1.25.0
packaging==23.1
pip-upgrader==1.4.15
platformdirs==3.8.1
pyaml==23.7.0
pylint==2.17.4
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0
requests==2.31.0
soco==0.29.1
terminaltables==3.1.10
tomlkit==0.11.8
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
xmltodict==0.13.0
